Fix permissions in user modal

+++ frontend/src/app/modules/organization/modules/users/modals/user/user.ts
@@ -6,9 +6,9 @@ import { IdentityService } from "app/services/identity";
 import { OrganizationRole, OrganizationService, OrganizationTeam, OrganizationUser } from "app/model/organization";
 import { addToRecentEntities } from "app/modules/starred/services/recent-item";
 
- /**
-  * Represents a modal panel that presents the public profile of a user.
-  */
+/**
+ * Represents a modal panel that presents the public profile of a user.
+ */
 @autoinject
 export class UserModalPanel
 {
@@ -45,6 +45,16 @@ export class UserModalPanel
     protected readonly: boolean;
 
     /**
+     * True if authorized to change the role, otherwise false.
+     */
+    protected canChangeRole: boolean;
+
+    /**
+     * True if authorized to change the team, otherwise false.
+     */
+    protected canChangeTeam: boolean;
+
+    /**
      * The user to present.
      */
     protected user: OrganizationUser;
@@ -57,7 +67,7 @@ export class UserModalPanel
     /**
      * The selected teams, if any.
      */
-    protected selectedTeams: OrganizationTeam[] | undefined;
+    protected selectedTeams: { id: string; name: string }[] | undefined;
 
     /**
      * The available roles.
@@ -67,7 +77,7 @@ export class UserModalPanel
     /**
      * The selected role.
      */
-    protected selectedRole: OrganizationRole | undefined;
+    protected selectedRole: { id: string; name: string } | undefined;
 
     /**
      * The validation for the modal.
@@ -95,12 +105,19 @@ export class UserModalPanel
 
     /**
      * Called by the framework when the modal is activating.
-     * @param user The user to present.
+     * @param model The model to use.
      */
     public activate(model: { readonly: boolean; user: OrganizationUser }): void
     {
         this.user = model.user;
         this.readonly = model.readonly ?? false;
+
+        this.canChangeRole =
+            this._identityService.identity!.claims.has("edit-users");
+
+        this.canChangeTeam =
+            this._identityService.identity!.claims.has("view-teams") &&
+            this._identityService.identity!.claims.has("edit-teams");
 
         this.operation = new Operation(async () =>
         {
@@ -110,12 +127,17 @@ export class UserModalPanel
             {
                 [this.availableRoles, this.availableTeams] = await Promise.all(
                 [
-                    await this._organizationService.getRoles(),
-                    await this._organizationService.getTeams()
+                    this.canChangeRole ? await this._organizationService.getRoles() : undefined,
+                    this.canChangeTeam ? await this._organizationService.getTeams() : undefined
                 ]);
 
-                this.selectedRole = this.availableRoles.find(r => r.id === this.user.role.id)!;
-                this.selectedTeams = this.availableTeams.filter(t1 => this.user.teams.some(t2 => t2.id === t1.id));
+                this.selectedRole = this.canChangeRole
+                    ? this.availableRoles!.find(r => r.id === this.user.role.id)!
+                    : this.user.role;
+
+                this.selectedTeams = this.canChangeTeam
+                    ? this.availableTeams!.filter(t1 => this.user.teams.some(t2 => t2.id === t1.id))
+                    : this.user.teams;
 
                 this._modal.busy = false;
             }
@@ -139,7 +161,7 @@ export class UserModalPanel
     /**
      * Called when the `Save changes` button is clicked.
      */
-    protected async onSubmitClick(): Promise<void>
+    protected async onSaveClick(): Promise<void>
     {
         try
         {
@@ -154,32 +176,38 @@ export class UserModalPanel
 
             this._modal.busy = true;
 
-            const updateTeamsPromises: Promise<any>[] = [];
-
-            for (const team of this.selectedTeams?.filter(t1 => !this.user.teams.some(t2 => t2.id === t1.id)) ?? [])
+            if (this.canChangeTeam)
             {
-                updateTeamsPromises.push(this._organizationService.addUserToTeam(team.id, this.user.id)
-                    .then(() => this.user.teams.push(team)));
+                const updateTeamsPromises: Promise<any>[] = [];
+
+                for (const team of this.selectedTeams?.filter(t1 => !this.user.teams.some(t2 => t2.id === t1.id)) ?? [])
+                {
+                    updateTeamsPromises.push(this._organizationService.addUserToTeam(team.id, this.user.id)
+                        .then(() => this.user.teams.push(team)));
+                }
+
+                for (const team of this.user.teams.filter(t1 => !this.selectedTeams?.some(t2 => t2.id === t1.id)) ?? [])
+                {
+                    updateTeamsPromises.push(this._organizationService.removeUserFromTeam(team.id, this.user.id)
+                        .then(() => this.user.teams.splice(this.user.teams.findIndex(t => t.id === team.id), 1)));
+                }
+
+                if (updateTeamsPromises.length > 0)
+                {
+                    // Ensure bindings are updated.
+                    this.user.teams = [...this.user.teams];
+                }
+
+                await Promise.all(updateTeamsPromises);
             }
 
-            for (const team of this.user.teams.filter(t1 => !this.selectedTeams?.some(t2 => t2.id === t1.id)) ?? [])
+            if (this.canChangeRole)
             {
-                updateTeamsPromises.push(this._organizationService.removeUserFromTeam(team.id, this.user.id)
-                    .then(() => this.user.teams.splice(this.user.teams.findIndex(t => t.id === team.id), 1)));
-            }
-
-            if (updateTeamsPromises.length > 0)
-            {
-                // Ensure bindings are updated.
-                this.user.teams = [...this.user.teams];
-            }
-
-            await Promise.all(updateTeamsPromises);
-
-            if (this.selectedRole!.id !== this.user.role.id)
-            {
-                await this._organizationService.changeUserRole(this.user.id, this.selectedRole!.id)
-                    .then(() => this.user.role = this.selectedRole!);
+                if (this.selectedRole!.id !== this.user.role.id)
+                {
+                    await this._organizationService.changeUserRole(this.user.id, this.selectedRole!.id)
+                        .then(() => this.user.role = this.selectedRole!);
+                }
             }
 
             if (this.user.id === this._identityService.identity!.id)
